reorganize column definitions
